fix: update tests to reflect Uint8Array
Puppeteer silently changed their API to return a Uint8Array instead of
a buffer causing test to fail.  Here, I am simply updating the tests.
If we wanted to keep the old API, we could call new Buffer() on the
output of the render function, but I believe this is likely unnecessary.